LPD-85583 Isolate the preview selectors and create a hook to consume de state
